Improve init package name guess
